refactor: change order to orders model, wich caused a conflict when creating the table in the DB

+++ backend/src/main/java/com/tap/serve/singapur/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.tap.serve.singapur.repository;
 import com.tap.serve.singapur.model.OrderModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<OrderModel, Integer> {
+public interface OrderRepository extends JpaRepository<OrderModel, Long> {
 }
